Move key inheritance utilities to their own file
